+++ config/routes.rb
@@ -8,6 +8,7 @@ Rails.application.routes.draw do
   root to: 'visitors#index'
   get 'resume', to: 'visitors#resume'
   get 'connect', to: 'visitors#connect'
+  get 'writing-corner', to: 'visitors#writing'
 
   devise_for :users, controllers: { registrations: 'users/registrations' }
   resources :users